backend: Make standalone the default production architecture.
