+++ packages/backend-api/package.json
@@ -12,11 +12,10 @@
     "start": "node ./server.js",
     "lint": "find src -name '*.ts' | xargs ../../node_modules/.bin/tslint -c ../../tslint.json",
     "lint:fix": "find src -name '*.ts' | xargs ../../node_modules/.bin/tslint -c ../../tslint.json --fix",
-    "watch": "PORT=8080 DEBUG_PORT=5859 npm run server",
     "compile": "rm -rf dist && ../../node_modules/.bin/tsc --sourceMap --outDir dist --declaration",
     "compile:watch": "../../node_modules/.bin/tsc --sourceMap --outDir dist --declaration --watch --preserveWatchOutput",
     "build": "npm run compile",
-    "server": "nodemon --delay 5 --inspect=5858 --legacy-watch --watch 'dist/**/*.js' --ignore 'test' ./server.js"
+    "watch": "nodemon --delay 5 --inspect=5859 --watch 'dist/**/*.js' --ignore 'test' dist/server.js"
   },
   "license": "Apache-2.0",
   "dependencies": {
