--- packages/server/README.md
@@ -1,5 +1,0 @@
-Create a test certificate via the following command
-
-```bash
-openssl req -x509 -newkey rsa:4096 -keyout key.pem -out cert.pem -days 365 -nodes
-```
